Refactor injection classes to separate package
