+++ src/main/java/tweeting/injection/modules/TwitterResourceModule.java
@@ -1,7 +1,8 @@
-package tweeting.resources;
+package tweeting.injection.modules;
 
 import dagger.Module;
 import dagger.Provides;
+import tweeting.resources.TwitterResource;
 import tweeting.services.TwitterService;
 
 import javax.inject.Singleton;
